+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - MaxPrioritySearchTree
   - A "dynamic" constructor option now allows deletion of the "top" (root) node. This is useful in certain algorithms.
 
+- Algorithms
+  - Implement the Maximal Empty Rectangle algorithm of De et al. It uses a dynamic MaxPST.
+
 ## [0.4.1] 2023-01-12
 
 - Update this file for the gem (though I forgot to add this comment first!)